shop/requirements.txt
asgiref==3.5.2
click==8.1.3
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
mypy-extensions==0.4.3
pathspec==0.10.3
Pillow==9.3.0
platformdirs==2.6.0
pytz==2022.6
sqlparse==0.4.3
